numpy v2: manually downcast from int64 to int32
seems that is what numpy <2 did silently but >2 raises an overflow error

+++ src/simplemseed/steimframeblock.py
@@ -162,7 +162,8 @@ class SteimFrameBlock:
             self.currentFrame += 1
 
         pos = self.currentSteimFrame.pos  # word position
-        self.currentSteimFrame.word[pos] = word  # add word
+        self.currentSteimFrame.word[pos] = word.astype(
+            self.currentSteimFrame.word.dtype)  # add word
         self.addEncodingNibble(nibble)  # add nibble
         self.numSamples += samples
         pos += 1  # increment position in frame
